requirements.txt
Pyrogram==2.0.30
TgCrypto==1.2.3
python-dotenv
aiohttp